Implementing install/publish. Update to follow design doc
Summary:
This diff will modify thrift-god into following the semantics
described in the thrift-registry RFC (http://t.uber.com/thrift-registry-rfc).
Modifications will include: (1) Addition of a publish command as an alternative
to the daemonized watcher. This publish command will be executed by a jenkins
build job, which will have been triggered by a Phabricator Herald rule that
triggers whenever a commit with changes to any *.thrift file in ./thrift/ lands
on master of any repo. (2) Addition of an install command. (3) Updating the
logic of the list command to reflect the new folder structure of the registry
and of the locally managed ./thrift/ folder.

Forgive the big initial commit. I've been wading through this code trying to
formulate a strategy for refactoring it for our purposes. Updating of tests
should show up in the next few commits.

Reviewers: jakev, kris

Subscribers: joshua, juncaoli, blampe, grayson, suurkivi, abg, prashant

Differential Revision: https://code.uberinternal.com/D188719